Fix fseek

+++ stock.c
@@ -30,7 +30,8 @@ void modifierProduit(const char* nomFichier) {
             printf("Nouvelle marque (actuelle: %s) : ", p.marque); scanf("%s", p.marque);
             printf("Nouveau prix (actuel: %.2f) : ", p.prix); scanf("%f", &p.prix);
             
-            fseek(file, -(long)sizeof(Produit), SEEK_CUR);            fwrite(&p, sizeof(Produit), 1, file);
+            fseek(file, -(long)sizeof(Produit), SEEK_CUR);
+            fwrite(&p, sizeof(Produit), 1, file);
             printf("Modification enregistree !\n");
             break;
         }
@@ -74,7 +75,7 @@ void vendreProduit(const char* nomFichier) {
             if (p.quantite >= qteVendre) {
                 p.quantite -= qteVendre;
                 // On remonte le curseur pour ecraser l'ancienne valeur
-                fseek(file, -sizeof(Produit), SEEK_CUR);
+                fseek(file, -(long)sizeof(Produit), SEEK_CUR);
                 fwrite(&p, sizeof(Produit), 1, file);
                 printf("Vente enregistree. Nouveau stock : %d\n", p.quantite);
             } else {
